requirements.txt
psutil
plotly
pandas
dash
dash_bootstrap_components
dash_bootstrap_template
